refactor(nx-cmake): move nx and platform utils to generator utils
