+++ packages/nx-cmake/src/utils/generatorUtils/checkNxVersion/checkNxVersion.ts
@@ -1,7 +1,7 @@
 import {
     REQUIRED_MAJOR_NX_VERSION,
     REQUIRED_MINOR_NX_VERSION,
-} from '../../../../config/version';
+} from '../../../config/version';
 
 export const checkNxVersion = (version: string): boolean => {
     const [major, minor, patch] = version
