+++ packages/nx-cmake/src/utils/generatorUtils/checkOs/checkOs.ts
