+++ packages/nx-cmake/src/utils/generatorUtils/getRequiredVersionOfNx/getRequiredVersionOfNx.spec.ts
